Updating accountID from int to int64

+++ pkg/dbt_cloud/athena_credential.go
@@ -41,7 +41,7 @@ type AthenaCredentialData struct {
 // It doesn't include the UnencryptedCredentialDetails field which is only returned by the API
 type AthenaCredentialRequest struct {
 	ID                *int                     `json:"id,omitempty"`
-	AccountID         int                      `json:"account_id"`
+	AccountID         int64                    `json:"account_id"`
 	ProjectID         int                      `json:"project_id"`
 	Type              string                   `json:"type"`
 	State             int                      `json:"state"`
